embed rpc.Clients in zetaclient/zetacored.Client
